split class TestDataFrameConcat to file test_dataframe.py
